{{cookiecutter.project_name}}/pyproject.toml
[project]
name = "{{cookiecutter.project_name}}"
dynamic = ["version"]
description = "{{cookiecutter.project_description}}"
readme = "README.md"
requires-python = ">={{cookiecutter.min_python_version}}"
dependencies = [
    "fastapi>=0.115.8",
    "google-api-python-client>=2.162.0",
    "google-cloud-build>=3.31.0",
    "google-cloud-iam>=2.18.1",
    "google-cloud-resource-manager>=1.14.1",
    "google-cloud-secret-manager>=2.23.1",
    "pygithub>=2.6.1",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["setuptools>=42", "setuptools_scm[toml]>=6.0.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.4",
    "faker>=36.1.1",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
]

[tool.basedpyright]
ignore = ['tests/main_test']
reportUninitializedInstanceVariable="none"